Added cbot_control package
